requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.6
